refactor: move TryFindRandomTileOnStation methods to new helper system
