+++ Content.Server/_Scp/Scp106/Systems/Scp106System.cs
@@ -1,21 +1,15 @@
-﻿using System.Diagnostics.CodeAnalysis;
-using System.Linq;
+﻿using System.Linq;
 using System.Threading.Tasks;
-using Content.Server.Atmos.EntitySystems;
+using Content.Server._Scp.Helpers;
 using Content.Server.Gateway.Systems;
-using Content.Server.Station.Components;
 using Content.Shared._Scp.Scp106.Components;
 using Content.Shared._Scp.Scp106.Systems;
 using Content.Shared.Humanoid;
 using Content.Shared.Mind;
 using Content.Shared.Mobs;
-using Content.Shared.Random.Helpers;
-using Robust.Server.GameObjects;
 using Robust.Shared.Audio;
 using Robust.Shared.Audio.Systems;
-using Robust.Shared.Collections;
 using Robust.Shared.Map;
-using Robust.Shared.Map.Components;
 using Robust.Shared.Random;
 
 namespace Content.Server._Scp.Scp106.Systems;
@@ -25,10 +19,9 @@ public sealed class Scp106System : SharedScp106System
     [Dependency] private readonly IRobustRandom _random = default!;
     [Dependency] private readonly StairsSystem _stairs = default!;
     [Dependency] private readonly SharedTransformSystem _transform = default!;
-    [Dependency] private readonly AtmosphereSystem _atmosphere = default!;
-    [Dependency] private readonly MapSystem _map = default!;
     [Dependency] private readonly SharedMindSystem _mindSystem = default!;
     [Dependency] private readonly SharedAudioSystem _audio = default!;
+    [Dependency] private readonly ScpHelpersSystem _scpHelpers = default!;
 
     private readonly SoundSpecifier _sendBackroomsSound = new SoundPathSpecifier("/Audio/_Scp/Scp106/onbackrooms.ogg");
 
@@ -91,7 +84,7 @@ public sealed class Scp106System : SharedScp106System
 
     public override void SendToStation(EntityUid target)
     {
-        if (!TryFindRandomTile(out _, out _, out _, out var targetCoords))
+        if (!_scpHelpers.TryFindRandomTile(out _, out _, out _, out var targetCoords))
             return;
 
         _transform.SetCoordinates(target, targetCoords);
@@ -114,102 +107,6 @@ public sealed class Scp106System : SharedScp106System
         return EntityQuery<Scp106BackRoomMarkComponent>()
             .Select(entity => Transform(entity.Owner).Coordinates)
             .ToHashSet();
-    }
-
-    private bool TryGetRandomStation([NotNullWhen(true)] out EntityUid? station, Func<EntityUid, bool>? filter = null)
-    {
-        var stations = new ValueList<EntityUid>(Count<StationEventEligibleComponent>());
-
-        filter ??= _ => true;
-        var query = AllEntityQuery<StationEventEligibleComponent>();
-
-        while (query.MoveNext(out var uid, out _))
-        {
-            if (!filter(uid))
-                continue;
-
-            stations.Add(uid);
-        }
-
-        if (stations.Count == 0)
-        {
-            station = null;
-            return false;
-        }
-
-        station = stations[_random.Next(stations.Count)];
-        return true;
-    }
-
-    public bool TryFindRandomTile(out Vector2i tile,
-        [NotNullWhen(true)] out EntityUid? targetStation,
-        out EntityUid targetGrid,
-        out EntityCoordinates targetCoords)
-    {
-        tile = default;
-        targetStation = EntityUid.Invalid;
-        targetGrid = EntityUid.Invalid;
-        targetCoords = EntityCoordinates.Invalid;
-        if (TryGetRandomStation(out targetStation))
-        {
-            return TryFindRandomTileOnStation((targetStation.Value, Comp<StationDataComponent>(targetStation.Value)),
-                out tile,
-                out targetGrid,
-                out targetCoords);
-        }
-
-        return false;
-    }
-
-    public bool TryFindRandomTileOnStation(Entity<StationDataComponent> station,
-        out Vector2i tile,
-        out EntityUid targetGrid,
-        out EntityCoordinates targetCoords)
-    {
-        tile = default;
-        targetCoords = EntityCoordinates.Invalid;
-        targetGrid = EntityUid.Invalid;
-
-        var weights = new Dictionary<Entity<MapGridComponent>, float>();
-        foreach (var possibleTarget in station.Comp.Grids)
-        {
-            if (!TryComp<MapGridComponent>(possibleTarget, out var comp))
-                continue;
-
-            weights.Add((possibleTarget, comp), _map.GetAllTiles(possibleTarget, comp).Count());
-        }
-
-        if (weights.Count == 0)
-        {
-            targetGrid = EntityUid.Invalid;
-            return false;
-        }
-
-        (targetGrid, var gridComp) = _random.Pick(weights);
-
-        var found = false;
-        var aabb = gridComp.LocalAABB;
-
-        for (var i = 0; i < 10; i++)
-        {
-            var randomX = _random.Next((int) aabb.Left, (int) aabb.Right);
-            var randomY = _random.Next((int) aabb.Bottom, (int) aabb.Top);
-
-            tile = new Vector2i(randomX, randomY);
-            if (_atmosphere.IsTileSpace(targetGrid, Transform(targetGrid).MapUid, tile)
-                || _atmosphere.IsTileAirBlocked(targetGrid, tile, mapGridComp: gridComp)
-                || !_map.TryGetTileRef(targetGrid, gridComp, tile, out var tileRef)
-                || tileRef.Tile.IsEmpty)
-            {
-                continue;
-            }
-
-            found = true;
-            targetCoords = _map.GridTileToLocal(targetGrid, gridComp, tile);
-            break;
-        }
-
-        return found;
     }
 
     public override void Update(float frameTime)
